Fix listing of tags
The problem with the current approach is that it only fetches a subset of all tags on the repo. Unfortunately it is not always guaranteed that the set fetched will include the latest tag. There’s also no way to fetch tags by creation time. Hence, this approach fetches all tags first, and then filters them